move types to types-folder, tests into test folder, update typings
