+++ .taprc
@@ -1,0 +1,2 @@
+files:
+  - test/**/*.test.js
